CRUD update

+++ base/views.py
@@ -1,7 +1,13 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
-from .models import Room, Topic
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth.forms import UserCreationForm 
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import login, authenticate, logout
+from .models import Room, Topic, Message
 from .forms import RoomForm
+from django.http import HttpResponse
 # Create your views here.
 
 # rooms = [
@@ -10,7 +16,55 @@ from .forms import RoomForm
 #     {'id':3, 'name':'Is it worth it to invest in Nvidia?'},
 # ]
 
-def home(request): 
+def loginpage(request): 
+    page = 'login'
+    # we don't want the user to access the login page again if they are already logged in 
+    if request.user.is_authenticated: 
+        return redirect('home')
+    
+    # note do not call login, as login is a built in function and calling/name it login may cause clashes 
+    if request.method == "POST":
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try: 
+            user = User.objects.get(username=username)
+        except: 
+            messages.error(request, 'User does not exist')
+    # we are making use of django flash messages, which are messages which are stored inside of django and stored in only 1 browser refresh 
+        user = authenticate(request, username=username, password=password)
+        # this is verifiying whether the user exists
+
+        if user is not None: 
+            login(request, user)
+            return redirect('home')
+        
+        else:
+            messages.error(request, 'Username or Password is incorrect, please try again.')
+
+    context = {'page':page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutuser(request): 
+    logout(request)
+    return redirect ('home')
+
+def registeruser(request): 
+    form = UserCreationForm()
+    if request.method == "POST": 
+        form = UserCreationForm(request.POST)
+        if form.is_valid(): 
+            user = form.save(commit=False)
+            # we want to be able to access the user right way hence we set commit to False, essentially freezing it in time
+            user.username = user.username
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else: 
+            messages.error(request, 'An error occured during registation') 
+    return render(request, 'base/login_register.html', {'form':form})
+
+def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
         Q(topic__name__icontains=q) |
@@ -25,14 +79,23 @@ def home(request):
 
 def room(request, pk): 
     roomval = Room.objects.get(id=pk)
-    # for i in rooms:
-    #     if i['id'] == int(pk):
-    #         roomval = i 
-    context = {'room': roomval}
+    room_messages = roomval.message_set.all().order_by('-created')
+    # message_set.all means that messages related to the parent owner (in this case room) is called 
+    if request.method == 'POST': 
+        message = Message.objects.create(
+            user=request.user,
+            room=roomval,
+            body=request.POST.get('body')
+        )
+        return redirect('room', pk=roomval.id)
+        # note it is "objects" not object and also note that pk is roomval.id not room.id 
+        # return redirect('home')
+    
+    context = {'room': roomval, 'room_messages': room_messages}
     return render(request, 'base/roompg.html', context)
 
 # note the pk in rooms which is also used in urls, we can use pk as an id as it will always be unique
-
+@login_required(login_url="login")
 def createroom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -45,10 +108,14 @@ def createroom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url="login")
 def updateroom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host: 
+        return HttpResponse('Room host not found')
+    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -58,8 +125,13 @@ def updateroom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url="login")
 def deleteroom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host: 
+        return HttpResponse('Room host not found')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
